pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11.8"
langchain = "^0.1.16"
fastapi = "^0.110.1"
langserve = {extras = ["all"], version = "^0.1.0"}
langchain-cli = "^0.0.21"
langchain-openai = "^0.1.3"
python-decouple = "^3.8"
sentence-transformers = "^2.7.0"
chromadb = "^0.4.24"
python-multipart = "^0.0.9"
pydantic = "1.10.13"
pypdf = "^4.2.0"
weaviate-client = "==3.*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
